Extract MarketPriceRequestService from MarketPriceService Apply changes to client code
