+++ rest_api_app/src/main/resources/rest_api.conf
@@ -36,10 +36,6 @@ application {
                 operator = "wiz",
             },
             {
-                url = "http://wizpriceje6q5tdrxkyiazsgu7irquiqjy2dptezqhrtu7l2qelqktid.onion/"
-                operator = "wiz",
-            },
-            {
                 url = "http://emzypricpidesmyqg2hc6dkwitqzaxrqnpkdg3ae2wef5znncu2ambqd.onion/"
                 operator = "emzy",
             },
