+++ seed_node_app/src/main/resources/seed_node.conf
@@ -19,10 +19,6 @@ application {
                 operator = "wiz",
             },
             {
-                url = "http://wizpriceje6q5tdrxkyiazsgu7irquiqjy2dptezqhrtu7l2qelqktid.onion/"
-                operator = "wiz",
-            },
-            {
                 url = "http://emzypricpidesmyqg2hc6dkwitqzaxrqnpkdg3ae2wef5znncu2ambqd.onion/"
                 operator = "emzy",
             },
